+++ vite.config.ts
@@ -6,6 +6,13 @@ import { dependencies, peerDependencies } from "./package.json";
 const DEPS_TO_BUNDLE = ["graphql-request", "graphql"];
 
 export default defineConfig({
+  optimizeDeps: {
+    esbuildOptions: {
+      define: {
+        global: "globalThis",
+      },
+    },
+  },
   plugins: [
     plugin({
       jsxRuntime: "classic",